back/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
